Finishing touches
Fixed variant.txt, added points for Bayes sets, reworked plotting to include legend for both lines and points, changed the get_coefs function to include the case when the value is -1, changed the get_points function to include the case when the value of the parameter b of the equation p_2 = k*p_1+b is 0 and to include all the different cases of parameters being 0 for point calculation, changed plotting_lines and create_new_row functions to build better graphs, wrote a report. Hope it's finally done.